code-refactoring after review
